requirements.txt
# Dependencies
Django>=4.2.8
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
django-redis>=5.4.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
djoser>=2.2.0
Pillow>=10.0.0
django-filter>=23.5
stripe>=7.0.0
python-decouple>=3.8
celery>=5.3.0
django-cors-headers>=4.3.0
requests>=2.31.0
pycountry>=23.12.0

# Code quality tools
flake8>=6.1.0
black>=23.11.0